+++ examples/lavalink-basic-bot.rs
@@ -47,6 +47,11 @@ async fn main() -> anyhow::Result<()> {
     // Initialize the tracing subscriber.
     tracing_subscriber::fmt::init();
 
+    // Select rustls backend
+    rustls::crypto::ring::default_provider()
+        .install_default()
+        .unwrap();
+
     let (mut shard, state) = {
         let token = env::var("DISCORD_TOKEN").expect("Missing DISCORD_TOKEN");
         let lavalink_host =